Fixed merge problems
