+++ src/Robot/Kicker.cpp
@@ -1,0 +1,29 @@
+// kicker
+
+/**
+//General Framework for the Kicker Code
+
+//Manual Wind Up Arm
+//Up On D-Pad
+void manualWindUp() {
+  
+}
+
+//Auto Wind Up
+//Triangle on Controller
+void autoWindUp() {
+
+}
+
+//Slow Release Arm
+//Down on DPad
+void slowRelease() {
+
+}
+
+//Fire
+//X on Controller
+void fireKicker() {
+
+}
+**/
